requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
requests==2.27.1
soupsieve==2.3.1
urllib3==1.26.8
opencv-python~=4.5.5.64
numpy~=1.22.3
PyYAML~=6.0
pyinstaller